% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_connectivity.R
\name{calc_influx}
\alias{calc_influx}
\title{Calculate influx}
\usage{
calc_influx(data, dest_col, N_col, ...)
}
\arguments{
\item{data}{Dataframe in long format}

\item{dest_col}{Destination site (unquoted column name)}

\item{N_col}{Lice density (unquoted column name)}

\item{...}{Grouping columns (unquoted)}
}
\value{
Summarised dataframe with column \code{total} giving the sum. Note
that this includes self infection.
}
\description{
Calculate influx
}
