% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_psteps.R
\name{calc_psteps_avg}
\alias{calc_psteps_avg}
\title{Calculate average particle densities}
\usage{
calc_psteps_avg(ps_long, y_col, grp_col = NULL, ncores = 4, mesh_sf = NULL)
}
\arguments{
\item{ps_long}{Long-format dataframe}

\item{y_col}{Quoted column name with particle densities}

\item{grp_col}{\code{NULL} or quoted column name with time to average over (e.g., years or months)}

\item{ncores}{Number of cores for parallelization with furrr}

\item{mesh_sf}{\code{NULL} or mesh sf object}
}
\description{
Calculate average particle densities
}
