% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_psteps.R
\name{calc_psteps_diff}
\alias{calc_psteps_diff}
\title{Calculate difference in particle densities between to simulations}
\usage{
calc_psteps_diff(ps_wide, sims_comp, ncores = 4)
}
\arguments{
\item{ps_wide}{Output from \code{load_psteps_simSets()} or similar, with
columns \code{i} with element indexes, \code{sim} with simulation
identifier, and columns \code{t_*} with particle densities per timestep
\code{t}}

\item{sims_comp}{Vector of length two indicating which simulations to
compare, corresponding to values in column \code{sim}}

\item{ncores}{Number of cores for parallel processing using the \code{furrr}
package}
}
\value{
Dataframe with one row per element in column \code{i} and columns
  \code{t_*} with differences. Note that the difference is \code{last() -
  first()}, with order by R default (i.e., \code{sort(sims_comp)}) and _NOT_
  the order given in \code{sims_comp}. Elements where \code{is.na(first())}
  return \code{9999}, elements where \code{is.na(last())} return
  \code{-9999}.
}
\description{
Calculate difference in particle densities between to simulations
}
