% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_connectivity.R
\name{calc_self_infection}
\alias{calc_self_infection}
\title{Calculate self infection}
\usage{
calc_self_infection(data, src_col, dest_col, N_col, ...)
}
\arguments{
\item{data}{Dataframe in long format}

\item{src_col}{Source site (unquoted column name)}

\item{dest_col}{Destination site (unquoted column name)}

\item{N_col}{Lice density (unquoted column name)}

\item{...}{Grouping columns (unquoted)}
}
\value{
Summarised dataframe with column \code{total} giving the sum
}
\description{
Calculate self infection
}
