% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_out_dir.R
\name{clean_out_dir}
\alias{clean_out_dir}
\title{Clean output directory}
\usage{
clean_out_dir(pattern, newdir)
}
\arguments{
\item{pattern}{File name pattern}

\item{newdir}{New directory name}
}
\value{
No return
}
\description{
Simulations put all files in the default output directory. This function moves file types to corresponding folders.
}
