% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_data.R
\name{download_fish_biomass}
\alias{download_fish_biomass}
\title{Download monthly Scottish fish farm biomass}
\usage{
download_fish_biomass(begin_ymd, end_ymd, out_file)
}
\arguments{
\item{begin_ymd}{Beginning date in YYYY-MM-DD format}

\item{end_ymd}{Ending date in YYYY-MM-DD format}

\item{out_file}{Filename for the csv output}
}
\value{
Success message if csv is correctly saved
}
\description{
Accessed through aquaculture.scotland.gov.uk. Downloads a csv of monthly fish
biomass on each farm.
}
\details{
http://aquaculture.scotland.gov.uk/data/data.aspx
}
