% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_data.R
\name{download_lice_counts}
\alias{download_lice_counts}
\title{Download weekly Scottish sea lice counts}
\usage{
download_lice_counts(begin_ymd, end_ymd, out_file)
}
\arguments{
\item{begin_ymd}{Beginning date in YYYY-MM-DD format}

\item{end_ymd}{Ending date in YYYY-MM-DD format}

\item{out_file}{Filename for the csv output}
}
\value{
Success message if csv is correctly saved
}
\description{
Accessed through aquaculture.scotland.gov.uk. Downloads a json file from the
ArcGIS online map viewer, then processes that into a csv.
}
\details{
Data from before 2021-March-29 must be downloaded manually from the website.
Data are the weekly average number of adult female Lepeophtheirus salmonis
reported by the salmon companies.

http://aquaculture.scotland.gov.uk/data/data.aspx
https://scottishepa.maps.arcgis.com/apps/webappviewer/index.html?id=2218824350e5470e8026076d4138da58
}
