% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_data.R
\name{download_ms_site_details}
\alias{download_ms_site_details}
\title{Download Marine Scotland site details}
\usage{
download_ms_site_details(out_file)
}
\arguments{
\item{out_file}{Filename for the csv output}
}
\value{
Success message if csv is correctly saved
}
\description{
http://aquaculture.scotland.gov.uk/data/data.aspx
http://aquaculture.scotland.gov.uk/data/site_details.aspx
}
