% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_data.R
\name{download_sepa_licenses}
\alias{download_sepa_licenses}
\title{Download SEPA farm license details}
\usage{
download_sepa_licenses(out_file)
}
\arguments{
\item{out_file}{Filename for the csv output}
}
\value{
Success message if csv is correctly saved
}
\description{
http://aquaculture.scotland.gov.uk/data/data.aspx
http://aquaculture.scotland.gov.uk/data/lease_details.aspx
}
