% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_psteps.R
\name{load_psteps}
\alias{load_psteps}
\title{Load psteps (particle densities per mesh element)}
\usage{
load_psteps(f, site_names = NULL)
}
\arguments{
\item{f}{Filename of psteps file output by biotracker}

\item{site_names}{Vector of site names; length(site_names) = (ncol(pstepsFile)-1)}
}
\value{
Dataframe with column \code{i} for mesh element, and colmns for particle densities named with either 'wk_' or the sitename plus the YYYYMMDD date (e.g., \code{wk_20190407}).
}
\description{
If biotracker is run with \code{splitPsteps='true'}, output is a dataframe with
column 'i' for mesh element and a column for each release site giving the
density within that file (e.g., weekly totals). In this case,
\code{site_names} must be provided. If \code{splitPsteps='false'}, biotracker
sums all particle densities within each element regardless of release site,
and \code{site_names} should be left \code{NULL}.
}
