% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_psteps.R
\name{load_psteps_simSets}
\alias{load_psteps_simSets}
\title{Load (sparse) particle densities for a set of simulations}
\usage{
load_psteps_simSets(
  out_dir,
  mesh_i,
  sim_i,
  ncores = 4,
  stage = "Mature",
  liceScale = 28.2 * 240,
  per_m2 = TRUE,
  log = TRUE
)
}
\arguments{
\item{out_dir}{Main output directory containing simulation-specific subdirectories}

\item{mesh_i}{Dataframe with mesh information including columns \code{i}=element and (optionally) \code{area}=element area}

\item{sim_i}{Simulation metadata with column \code{sim} naming each simulation. These should also be the subdirectory names}

\item{ncores}{Number of cores to use for parallel processing}

\item{stage}{Life stage ("Mature" or "Immature")}

\item{liceScale}{Multiplier for original pstep values}

\item{per_m2}{Logical: Scale values by element area?}

\item{log}{Logical: ln(values)? Performed after area scaling}
}
\value{
Wide format dataframe
}
\description{
Particle steps from simulations organized into subdirectories within
\code{out_dir} are loaded in wide format, with a column for each week and a
row for each mesh element x simulation (sparse: only elements with particles).
}
