% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_psteps.R
\name{load_vertDistr}
\alias{load_vertDistr}
\title{Load (sparse) vertical distributions in each element}
\usage{
load_vertDistr(f, liceScale = 28.2 * 240)
}
\arguments{
\item{f}{Filename of the vertDistr* file output by biotracker}

\item{liceScale}{Multiplier for the original density values}
}
\value{
Dataframe with column \code{i} for mesh element, \code{z} for depth bin, \code{hour} for the simulation hour, and \code{value} with the particle density.
}
\description{
Load (sparse) vertical distributions in each element
}
