% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_psteps.R
\name{load_vertDistr_simSets}
\alias{load_vertDistr_simSets}
\title{Load (sparse) vertical distributions for a set of simulations}
\usage{
load_vertDistr_simSets(
  out_dir,
  mesh_i,
  sim_i,
  ncores = 4,
  stage = "Mature",
  liceScale = 28.2 * 240,
  per_m2 = FALSE,
  trans = NULL,
  date_rng
)
}
\arguments{
\item{out_dir}{Main output directory containing simulation-specific subdirectories}

\item{mesh_i}{Dataframe with mesh information including columns \code{i}=element and (optionally) \code{area}=element area}

\item{sim_i}{Simulation metadata with column \code{sim} naming each simulation. These should also be the subdirectory names}

\item{ncores}{Number of cores to use for parallel processing}

\item{stage}{Life stage ("Mature" or "Immature")}

\item{liceScale}{Multiplier for original pstep values}

\item{per_m2}{Logical: Scale values by element area?}

\item{trans}{Transform values? Options are NULL, "log", or "4th_rt". Performed after area scaling}

\item{date_rng}{Date range; vector with min and max dates in YYYY-MM-DD format}
}
\value{
Dataframe
}
\description{
Load (sparse) vertical distributions for a set of simulations
}
