% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_biotracker.R
\name{run_biotracker}
\alias{run_biotracker}
\title{Run biotracker}
\usage{
run_biotracker(
  jdk_path = "/home/sa04ts/.jdks/jdk-23.0.1/bin/java",
  jar_path = "/home/sa04ts/biotracker/jar/biotracker.jar",
  f_properties = "sim_01.properties",
  sim_dir = "sim_01/",
  clean_output = TRUE,
  tar_output = FALSE,
  save_tar_only = FALSE
)
}
\arguments{
\item{jdk_path}{Path to java jdk}

\item{jar_path}{Path to biotracker.jar}

\item{f_properties}{Properties file}

\item{sim_dir}{Simulation directory}

\item{clean_output}{Move files to appropriate subdirectories?}

\item{tar_output}{Create tar.gz of all output?}

\item{save_tar_only}{Remove uncompressed folder and keep only tar.gz?}
}
\value{
Reporting only
}
\description{
Run biotracker
}
