% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_connectivity.R
\name{calc_outflux}
\alias{calc_outflux}
\title{Calculate outflux}
\usage{
calc_outflux(data, src_col, N_col, ...)
}
\arguments{
\item{data}{Dataframe in long format}

\item{src_col}{Source site (unquoted column name)}

\item{N_col}{Lice density (unquoted column name)}

\item{...}{Grouping columns (unquoted)}
}
\value{
Summarised dataframe with column \code{outflux} giving the sum. Note
that this includes self infection.
}
\description{
Calculate outflux
}
