% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_data.R
\name{download_lice_counts}
\alias{download_lice_counts}
\title{Download weekly Scottish sea lice counts}
\usage{
download_lice_counts(begin_ymd, end_ymd, out_file)
}
\arguments{
\item{begin_ymd}{Beginning date in YYYY-MM-DD format}

\item{end_ymd}{Ending date in YYYY-MM-DD format}

\item{out_file}{Filename for the csv output}
}
\value{
Success message if csv is correctly saved
}
\description{
Accessed through aquaculture.scotland.gov.uk. Downloads a json file from the
ArcGIS online map viewer, then processes that into a csv. Note that only data
from March 2021 are available through this function. Older data must be
downloaded manually from the website. Data are the weekly average number of
adult female Lepeophtheirus salmonis reported by the salmon companies.
}
