% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_connectivity.R
\name{load_connectivity}
\alias{load_connectivity}
\title{Load pairwise connectivity matrix}
\usage{
load_connectivity(f, site_names, liceScale = 28.2 * 240)
}
\arguments{
\item{f}{Filename of connectivity matrix output by biotracker}

\item{site_names}{Vector of site names; order corresponds to integers in connectivity csvs}

\item{liceScale}{Multiplier for original particle density values}
}
\value{
Dataframe with a column for source, destination, particle density transferred, and an identifier column with the date
}
\description{
Load pairwise connectivity matrix
}
