% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_connectivity.R
\name{load_connectivity}
\alias{load_connectivity}
\title{Load pairwise connectivity matrices}
\usage{
load_connectivity(f, site_names)
}
\arguments{
\item{f}{Vector of filenames}

\item{site_names}{Vector of site names; length(site_names) = ncol(connect_mx)}
}
\value{
Dataframe with a column for each site (destinations), and rows for each site (source) x date combination
}
\description{
Load pairwise connectivity matrices
}
