% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_biotracker.R
\name{run_biotracker}
\alias{run_biotracker}
\title{Run biotracker}
\usage{
run_biotracker(
  jdk_path = "/usr/local/java/jre1.8.0_211/bin/java",
  javaLib_path = "jar/lib/",
  jar_path = "/home/sa04ts/biotracker/jar/biotracker.jar",
  f_properties = "sim_01.properties",
  sim_dir = "sim_01/",
  clean_output = TRUE,
  tar_output = TRUE
)
}
\arguments{
\item{jdk_path}{Path to java jdk}

\item{javaLib_path}{Path to commons-io-2.6.jar and netcdfAll-4.6.11.jar}

\item{jar_path}{Path to biotracker.jar}

\item{f_properties}{Properties file}

\item{sim_dir}{Simulation directory}

\item{clean_output}{Move files to appropriate subdirectories?}

\item{tar_output}{Create tar.gz of all output?}
}
\value{
Reporting only
}
\description{
Run biotracker
}
