% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_biotracker_properties.R
\name{set_biotracker_properties}
\alias{set_biotracker_properties}
\title{Generate properties for biotracker}
\usage{
set_biotracker_properties(
  properties_file_path = NULL,
  coordRef = "OSGB1936",
  datadir = "/media/archiver/common/sa01da-work/WeStCOMS2/Archive/",
  datadirPrefix = "netcdf_",
  datadirSuffix = "",
  mesh1 = "/home/sa04ts/FVCOM_meshes/WeStCOMS2_mesh.nc",
  mesh1Type = "FVCOM",
  location = "westcoms",
  minchVersion = 2,
  sitefile = "../../data/farm_sites.tsv",
  siteDensityPath = "../../data/lice_daily_2019-04-01_2023-09-30.tsv",
  verboseSetUp = "false",
  start_ymd = 20190401,
  numberOfDays = 7,
  checkOpenBoundaries = "false",
  openBoundaryThresh = 500,
  duplicateLastDay = "true",
  recordsPerFile1 = 25,
  dt = 3600,
  verticalDynamics = "true",
  maxDepth = 10000,
  parallelThreads = 4,
  releaseScenario = 1,
  releaseInterval = 1,
  nparts = 1,
  setStartDepth = "true",
  startDepth = 1,
  fixDepth = "false",
  stepsPerStep = 30,
  variableDiffusion = "false",
  D_h = 0.1,
  D_hVert = 0.001,
  salinityThreshMin = 23,
  salinityThreshMax = 31,
  mortalityRate = 0.01,
  salinityMort = "true",
  swimLightLevel = "true",
  vertSwimSpeedMean = -5e-04,
  vertSwimSpeedStd = 1e-04,
  sinkingRateMean = 5e-04,
  sinkingRateStd = 1e-04,
  viabletime = -1,
  maxParticleAge = -1,
  viableDegreeDays = 40,
  maxDegreeDays = 150,
  recordPsteps = "false",
  splitPsteps = "false",
  pstepsInterval = 168,
  recordMovement = "false",
  recordElemActivity = "true",
  recordConnectivity = "true",
  connectivityInterval = 24,
  recordLocations = "false",
  recordArrivals = "false"
)
}
\arguments{
\item{properties_file_path}{filename for properties file; if NULL, character vector is returned only}

\item{coordRef}{Coordinate reference system}

\item{datadir}{Primary data directory}

\item{datadirPrefix}{Primary data file prefix}

\item{datadirSuffix}{Primary data file suffix}

\item{mesh1}{Primary mesh file}

\item{mesh1Type}{Primary mesh type}

\item{location}{Primary domain}

\item{minchVersion}{Primary mesh WeStCOMS version}

\item{sitefile}{Site location file (tsv)}

\item{siteDensityPath}{Site lice density path}

\item{verboseSetUp}{Additional info}

\item{start_ymd}{Start date YYYYMMDD}

\item{numberOfDays}{Number of days to simulate}

\item{checkOpenBoundaries}{Check for exits through open boundaries?}

\item{openBoundaryThresh}{Threshold for boundary exits (m)}

\item{duplicateLastDay}{TRUE; duplicates final day of hydrodynamic files}

\item{recordsPerFile1}{Records per hydrdynamic file}

\item{dt}{Time step}

\item{verticalDynamics}{Include vertical dynamics?}

\item{maxDepth}{Limit for particle depth (m, positive = deeper)}

\item{parallelThreads}{Number of threads}

\item{releaseScenario}{1 = consistent release across time steps}

\item{releaseInterval}{Interval (hours) for releaseScenario==1}

\item{nparts}{Particles per site per release}

\item{setStartDepth}{Set particle starting depth?}

\item{startDepth}{Particle starting depth}

\item{stepsPerStep}{RK4 steps per timestep}

\item{variableDiffusion}{Include variable vertical diffusion?}

\item{D_h}{Horizontal diffusion coefficient}

\item{D_hVert}{Vertical diffusion coefficient}

\item{salinityThreshMin}{Lower salinity threshold for sinking (all sink)}

\item{salinityThreshMax}{Upper salinity threshold for sinking (none sink)}

\item{mortalityRate}{Mortality rate if salinityMort=false}

\item{salinityMort}{Calculate mortality based on salinity?}

\item{swimLightLevel}{Should particles swim upward if light is sufficient?}

\item{vertSwimSpeedMean}{Mean upward swim speed (m/s)}

\item{vertSwimSpeedStd}{SD for upward swim speed (m/s)}

\item{sinkingRateMean}{Mean downward sinking speed (m/s)}

\item{sinkingRateStd}{SD for downward sinking speed (m/s)}

\item{viabletime}{Time at which particles become viable (h)}

\item{maxParticleAge}{Time at which particles die (h)}

\item{viableDegreeDays}{Degree days at which particles become viable}

\item{maxDegreeDays}{Degree days at which particles die}

\item{recordPsteps}{Record particle steps? Matrix with density/element}

\item{splitPsteps}{Split Psteps by release location?}

\item{pstepsInterval}{Psteps recording frequency (h)}

\item{recordMovement}{Record movement details for sample of particles?}

\item{recordElemActivity}{Record overall element/hourly particle activity?}

\item{recordConnectivity}{Record connectivity between sites?}

\item{connectivityInterval}{Connectivity recording frequency (h)}

\item{recordLocations}{Record hourly particle locations?}

\item{recordArrivals}{Record particle arrivals to sites?}

\item{destinationDirectory}{Output directory}

\item{datadir2}{Secondary data directory}

\item{datadir2Prefix}{Secondary data file prefix}

\item{datadir2Suffix}{Secondary data file suffix}

\item{mesh2}{Secondary mesh file}

\item{mesh2Type}{Secondary mesh type}

\item{location2}{Secondary domain}

\item{minchVersion2}{Secondary mesh WeStCOMS version}
}
\value{
Character vector with all properties; if properties_file_path is set, output is written to that file (overwriting any existing file)
}
\description{
Generate properties for biotracker
}
